refactor signing logic
